Remove agent_v2.py and update agent documentation
Delete the legacy agent_v2.py implementation. Update agent.py docstring
to describe the proxy workflow more precisely. Add _archive to
.gitignore and fix minor code formatting issues.
