+++ .gitignore
@@ -7,3 +7,4 @@ __pycache__/
 mempal.yaml
 .a5c/
 my-agent*
+_archive
